Fix ID int to long

+++ src/main/java/ru/aurorahunters/bot/service/solarwind/SunWindService.java
@@ -20,7 +20,7 @@ public class SunWindService implements Runnable {
     private String plasmaJson;
     private final int id;
 
-    public SunWindService(int id) {
+    public SunWindService(long id) {
         this.id = id;
     }
